requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==5.0.0a0
django-environ==0.12.0
django-extensions==3.2.3
environ==1.0
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
